requirements.txt
pyyaml
pytest
requests
coverage